create 404 page, update admin dashboard data & product stock management

+++ app/Http/Controllers/StripeController.php
@@ -40,7 +40,8 @@ class StripeController extends Controller
         if(Session::has('coupon')){
             $total_amount = Session::get('coupon')['total_amount'];
         }else{
-            $total_amount = round(Cart::total());
+            // $total_amount = round(Cart::total());
+            $total_amount = round((float) Cart::total());
         }
         
         $order_id = Order::insertGetId([
